refactor: split model catalog auth proof
