+++ src/agents/models-config.providers.auth-provenance.test.ts
@@ -754,135 +754,6 @@ describe("models-config provider auth provenance", () => {
     });
   });
 
-  it("uses canonical profile order for provider catalog discovery", () => {
-    const profileA = "openai:profile-a";
-    const profileB = "openai:profile-b";
-    const store: AuthProfileStore = {
-      version: 1,
-      profiles: {
-        [profileA]: {
-          type: "api_key",
-          provider: "openai",
-          key: "key-a",
-        },
-        [profileB]: {
-          type: "api_key",
-          provider: "openai",
-          key: "key-b",
-        },
-      },
-    };
-    const config: OpenClawConfig = {
-      auth: {
-        order: {
-          openai: [profileB, profileA],
-        },
-      },
-    };
-
-    expect(createProviderAuthResolver({}, store, config)("openai")).toMatchObject({
-      apiKey: "key-b",
-      profileId: profileB,
-    });
-    expect(createProviderApiKeyResolver({}, store, config)("openai")).toMatchObject({
-      apiKey: "key-b",
-      profileId: profileB,
-    });
-
-    store.order = { openai: [profileA, profileB] };
-    expect(createProviderAuthResolver({}, store, config)("openai")).toMatchObject({
-      apiKey: "key-a",
-      profileId: profileA,
-    });
-
-    delete store.order;
-    store.usageStats = {
-      [profileA]: {
-        cooldownUntil: Date.now() + 60_000,
-        cooldownReason: "rate_limit",
-        cooldownModel: "gpt-5.5",
-      },
-    };
-    const cooldownConfig: OpenClawConfig = {
-      auth: {
-        order: {
-          openai: [profileA, profileB],
-        },
-      },
-    };
-    expect(createProviderAuthResolver({}, store, cooldownConfig)("openai")).toMatchObject({
-      apiKey: "key-a",
-      profileId: profileA,
-    });
-    expect(createProviderApiKeyResolver({}, store, cooldownConfig)("openai")).toMatchObject({
-      apiKey: "key-a",
-      profileId: profileA,
-    });
-
-    store.usageStats = {
-      [profileA]: {
-        cooldownUntil: Date.now() + 60_000,
-        cooldownReason: "rate_limit",
-      },
-    };
-    expect(createProviderAuthResolver({}, store, cooldownConfig)("openai")).toMatchObject({
-      apiKey: "key-b",
-      profileId: profileB,
-    });
-    expect(createProviderApiKeyResolver({}, store, cooldownConfig)("openai")).toMatchObject({
-      apiKey: "key-b",
-      profileId: profileB,
-    });
-
-    const aliasConfig: OpenClawConfig = {
-      auth: {
-        order: {
-          "proof-alias": [profileB, profileA],
-        },
-      },
-    };
-    expect(createProviderAuthResolver({}, store, aliasConfig)("proof-alias")).toMatchObject({
-      apiKey: "key-b",
-      profileId: profileB,
-    });
-    expect(createProviderApiKeyResolver({}, store, aliasConfig)("proof-alias")).toMatchObject({
-      apiKey: "key-b",
-      profileId: profileB,
-    });
-  });
-
-  it("keeps unresolved OAuth refs selected for locked catalog resolution", () => {
-    const profileId = "openai:oauth-ref";
-    const auth = createProviderAuthResolver(
-      {},
-      {
-        version: 1,
-        profiles: {
-          [profileId]: {
-            type: "oauth",
-            provider: "openai",
-            access: "",
-            refresh: "",
-            expires: 0,
-            oauthRef: {
-              source: "openclaw-credentials",
-              provider: "openai-codex",
-              id: "00000000000000000000000000000000",
-            },
-          },
-        },
-      },
-      { auth: { order: { openai: [profileId] } } },
-    )("openai");
-
-    expect(auth).toMatchObject({
-      apiKey: undefined,
-      mode: "oauth",
-      profileId,
-      source: "profile",
-    });
-  });
-
   it("resolves plugin-owned synthetic auth through the provider hook", () => {
     // Plugin-owned synthetic auth can provide discovery keys while persisted
     // config still records a non-secret marker.
